Version 0.0.2: Added external Id properties in OutputProperty, refactored setting of OutputKeys in BaseOutputFormatter (some code moved into OutputTypeConfiguration), created CsvFileOutputFormatter

+++ src/main/java/org/planit/data/TraditionalStaticAssignmentSimulationData.java
@@ -124,7 +124,7 @@ public class TraditionalStaticAssignmentSimulationData extends SimulationData {
         return modalNetworkSegmentCostsMap.get(mode);
     }
     
-    public void setSkimMatrixValue(Mode mode, int originZone, int destinationZone, double tripCost) {
+    public void setSkimMatrixValue(Mode mode, long originZone, long destinationZone, double tripCost) {
     	modalSkimMatrixMap.get(mode).setValue(originZone, destinationZone, tripCost);
     }
