+++ src/main/java/org/planit/output/property/LinkSegmentExternalIdOutputProperty.java
@@ -19,7 +19,7 @@ public class LinkSegmentExternalIdOutputProperty extends BaseOutputProperty {
 
 	@Override
 	public Type getType() {
-		return Type.INTEGER;
+		return Type.LONG;
 	}
 
 	@Override
