+++ src/main/java/org/planit/output/property/ModeExternalIdOutputProperty.java
@@ -19,7 +19,7 @@ public final class ModeExternalIdOutputProperty extends BaseOutputProperty {
 
 	@Override
 	public Type getType() {
-		return Type.INTEGER;
+		return Type.LONG;
 	}
 
 	@Override
