+++ src/main/java/org/planit/output/property/RunIdOutputProperty.java
@@ -19,7 +19,7 @@ public final class RunIdOutputProperty extends BaseOutputProperty {
 
 	@Override
 	public Type getType() {
-		return Type.INTEGER;
+		return Type.LONG;
 	}
 
 	@Override
